+++ CHANGELOG.md
@@ -66,6 +66,10 @@
   the generated classes are `final` and thus the `protected` access modifier is equivalent to package access.
 * **\[processor\]** Omit the `public` access modifier on the constructor of generated component classes if
   the component class is not `public`.
+* **\[processor\]** Add the `public` access modifier to a generated component classes if it has a `@Inverse`
+  annotated method that has a target type in a different package and the component class does not have a public
+  constructor. This is required so that the generated classes can interact with the synthetic methods added
+  to component class to manage the inverse.
 
 ### [v0.104](https://github.com/arez/arez/tree/v0.104) (2018-08-01)
 [Full Changelog](https://github.com/arez/arez/compare/v0.103...v0.104)